Issue #2941494 by jhedstrom, Lendude: Deprecate SystemConfigFormTestBase and create kernel test version
